engines/requirements.txt
git+https://github.com/brainboost/EdgeGPT.git
git+https://github.com/brainboost/ChatGPT.git
git+https://github.com/brainboost/BingImageCreator.git
bardapi
google-generativeai
deepl
websockets
requests
curl_cffi
boto3
pillow